+++ src/components/app/App/index.js
@@ -22,6 +22,7 @@ import InProgressContainer from '../../../containers/columns/InProgressContainer
 import DoneContainer from '../../../containers/columns/DoneContainer';
 import ModalEditContainer from '../../../containers/modals/ModalEditContainer';
 import LoadingIndicator from '../../../components/feedback/LoadingIndicator';
+import NavigationContainer from '../../../containers/navigation/NavigationContainer';
 
 const App = ( { appStateFetched } ) => (
     <div className="container">
@@ -31,6 +32,7 @@ const App = ( { appStateFetched } ) => (
         <ToDoContainer />
         <InProgressContainer />
         <DoneContainer />
+        <NavigationContainer />
     </div>
 );