+++ src/containers/columns/BestIntentionsContainer/index.js
@@ -15,12 +15,11 @@
 
 import BestIntentions from '../../../components/columns/BestIntentions';
 
-import * as actions from './actions';
 import { connect } from 'kink';
 
 export default connect(
     BestIntentions,
-    actions,
+    {},
     ( { bestIntentions } ) => ( {
         items: ( bestIntentions ? bestIntentions.toJSON() : {} )
     } )