+++ src/lib/io/firebase/index.js
@@ -1,24 +1,28 @@
-/*
- * ▬▬ι═══════ﺤJediFocus-═══════ι▬▬
- *   use the source to get things done
- *
- * This project is a part of the “Byte-Sized JavaScript” videocasts.
- *
- * You can watch “Byte-Sized JavaScript” at: https://bit.ly/bytesized
- *
- * MIT Licensed — See LICENSE.md
- *
- * Send your comments, suggestions, and feedback to me@volkan.io                                                     l
+/*   __
+ *  |::|    Jedi Focus
+ *  |::|
+ * _|//|_  Do. Or do not.
+ * |____|  There is no try.
+ *  || |
+ *  | ||   This project is a part of the “Byte-Sized JavaScript” videocasts.
+ *  | ||   You can watch “Byte-Sized JavaScript” at: https://bytesized.tv/
+ *  || |
+ *  || |   MIT Licensed — See LICENSE.md
+ *  || |
+ *  | ||   Send your comments, suggestions, and feedback to me@volkan.io
+ *  ||||
  */
 
-const initialize = ( config ) => window.firebase.initializeApp( config );
+const db = () => window.firebase;
 
-const login = ( email, password ) => window.firebase.auth()
+const initialize = ( config ) => db().initializeApp( config );
+
+const login = ( email, password ) => db().auth()
     .signInWithEmailAndPassword( email, password );
 
 const ref = ( userId, context, path ) => `${userId}/${context}/${path}`;
 
-const refOnce = ( userId, context, path ) => window.firebase.database()
+const refOnce = ( userId, context, path ) => db().database()
     .ref( ref( userId, context, path ) ).once( 'value' )
     .then( ( snapshot ) => snapshot.val() );
 
@@ -28,7 +32,7 @@ const toDo = ( userId, context ) => refOnce( userId, context, 'toDo' );
 
 const inProgress = ( userId, context ) => refOnce( userId, context, 'inProgress' );
 
-const saveItem = ( userId, context, path, data ) => window.firebase.database()
+const saveItem = ( userId, context, path, data ) => db().database()
     .ref( ref( userId, context, path ) )
     .set( data );